Choose tile scaling in a set of predefined one's

+++ src/main/java/fr/thesmyler/terramap/proxy/TerramapProxy.java
@@ -13,6 +13,6 @@ public abstract class TerramapProxy {
 	public abstract void onServerHello(S2CTerramapHelloPacket pkt);
 	public abstract void onPlayerLoggedIn(PlayerLoggedInEvent event);
 	public abstract void onPlayerLoggedOut(PlayerLoggedOutEvent event);
-	public abstract float getDefaultGuiSize();
+	public abstract double getDefaultGuiSize();
 	
 }
